Toggle buttons Van Krevelen diagram
