+++ src/main/java/net/sf/mzmine/modules/visualization/kendrickmassplot/KendrickMassPlotParameters.java
@@ -41,70 +41,65 @@ import net.sf.mzmine.util.PeakListRowSorter;
 import net.sf.mzmine.util.SortingDirection;
 import net.sf.mzmine.util.SortingProperty;
 
+
+/**
+ * parameters for Kendrick mass plots
+ * 
+ * @author Ansgar Korf (ansgar.korf@uni-muenster.de)
+ */
 public class KendrickMassPlotParameters extends SimpleParameterSet {
-    public static final String rawDataFilesOption = "Raw data file";
+	public static final String rawDataFilesOption = "Raw data file";
 
-    public static final PeakListsParameter peakList = new PeakListsParameter(1,
-            1);
+	public static final PeakListsParameter peakList = new PeakListsParameter(1, 1);
 
-    public static final RawDataFilesParameter rawFile = new RawDataFilesParameter();
+	public static final RawDataFilesParameter rawFile = new RawDataFilesParameter();
 
-    public static final PeakSelectionParameter selectedRows = new PeakSelectionParameter();
+	public static final PeakSelectionParameter selectedRows = new PeakSelectionParameter();
 
-    public static final ComboParameter<String> yAxisValues = new ComboParameter<>(
-            "Y-Axis", "Select the kendrick mass defect base",
-            new String[] { "KMD (CH2)", "KMD (H)", "KMD (O)" });
+	public static final ComboParameter<String> yAxisValues = new ComboParameter<>("Y-Axis",
+			"Select the kendrick mass defect base", new String[] { "KMD (CH2)", "KMD (H)", "KMD (O)" });
 
-    public static final ComboParameter<String> xAxisValues = new ComboParameter<>(
-            "X-Axis",
-            "Select a second kendrick mass defect base, kendrick masse (KM) or m/z",
-            new String[] { "m/z", "KM", "KMD (CH2)", "KMD (H)", "KMD (O)" });
+	public static final ComboParameter<String> xAxisValues = new ComboParameter<>("X-Axis",
+			"Select a second kendrick mass defect base, kendrick masse (KM) or m/z",
+			new String[] { "m/z", "KM", "KMD (CH2)", "KMD (H)", "KMD (O)" });
 
-    public static final ComboParameter<String> zAxisValues = new ComboParameter<>(
-            "Z-Axis",
-            "Select a parameter for a third dimension, displayed as a heatmap or select none for a 2D plot",
-            new String[] { "none", "Retention time", "Intensity", "Area",
-                    "Tailing factor", "Asymmetry factor", "FWHM", "KMD (CH2)",
-                    "KMD (H)", "KMD (O)", "m/z" });
+	public static final ComboParameter<String> zAxisValues = new ComboParameter<>("Z-Axis",
+			"Select a parameter for a third dimension, displayed as a heatmap or select none for a 2D plot",
+			new String[] { "none", "Retention time", "Intensity", "Area", "Tailing factor", "Asymmetry factor", "FWHM",
+					"KMD (CH2)", "KMD (H)", "KMD (O)", "m/z" });
 
-    public static final ComboParameter<String> zScaleType = new ComboParameter<>(
-            "Z-Axis scale", "Select Z-Axis scale",
-            new String[] { "percentile", "custom" });
+	public static final ComboParameter<String> zScaleType = new ComboParameter<>("Z-Axis scale", "Select Z-Axis scale",
+			new String[] { "percentile", "custom" });
 
-    public static final DoubleRangeParameter zScaleRange = new DoubleRangeParameter(
-            "Range for z-Axis scale",
-            "Set the range for z-Axis scale."
-                    + " If percentile is used for z-Axis scale type, you can remove extreme values of the scale."
-                    + " E. g. type 0.5 and 99.5 to ignore the 0.5 smallest and 0.5 highest values. "
-                    + "If you choose custom, set ranges manually "
-                    + "Features out of scale range are displayed in magenta",
-            new DecimalFormat("##0.00"));
+	public static final DoubleRangeParameter zScaleRange = new DoubleRangeParameter("Range for z-Axis scale",
+			"Set the range for z-Axis scale."
+					+ " If percentile is used for z-Axis scale type, you can remove extreme values of the scale."
+					+ " E. g. type 0.5 and 99.5 to ignore the 0.5 smallest and 0.5 highest values. "
+					+ "If you choose custom, set ranges manually "
+					+ "Features out of scale range are displayed in magenta",
+			new DecimalFormat("##0.00"));
 
-    public static final ComboParameter<String> paintScale = new ComboParameter<>(
-            "Heatmap style", "Select the style for the third dimension",
-            new String[] { "Rainbow", "Monochrome red", "Monochrome green",
-                    "Monochrome yellow", "Monochrome cyan" });
+	public static final ComboParameter<String> paintScale = new ComboParameter<>("Heatmap style",
+			"Select the style for the third dimension",
+			new String[] { "Rainbow", "Monochrome red", "Monochrome green", "Monochrome yellow", "Monochrome cyan" });
 
-    public static final WindowSettingsParameter windowSettings = new WindowSettingsParameter();
+	public static final WindowSettingsParameter windowSettings = new WindowSettingsParameter();
 
-    public KendrickMassPlotParameters() {
-        super(new Parameter[] { peakList, rawFile, selectedRows, yAxisValues,
-                xAxisValues, zAxisValues, zScaleType, zScaleRange, paintScale,
-                windowSettings });
-    }
+	public KendrickMassPlotParameters() {
+		super(new Parameter[] { peakList, rawFile, selectedRows, yAxisValues, xAxisValues, zAxisValues, zScaleType,
+				zScaleRange, paintScale, windowSettings });
+	}
 
-    @Override
-    public ExitCode showSetupDialog(Window parent, boolean valueCheckRequired) {
+	@Override
+	public ExitCode showSetupDialog(Window parent, boolean valueCheckRequired) {
 
-        PeakList selectedPeakLists[] = getParameter(peakList).getValue()
-                .getMatchingPeakLists();
-        if (selectedPeakLists.length > 0) {
-            PeakListRow plRows[] = selectedPeakLists[0].getRows();
-            Arrays.sort(plRows, new PeakListRowSorter(SortingProperty.MZ,
-                    SortingDirection.Ascending));
-        }
+		PeakList selectedPeakLists[] = getParameter(peakList).getValue().getMatchingPeakLists();
+		if (selectedPeakLists.length > 0) {
+			PeakListRow plRows[] = selectedPeakLists[0].getRows();
+			Arrays.sort(plRows, new PeakListRowSorter(SortingProperty.MZ, SortingDirection.Ascending));
+		}
 
-        return super.showSetupDialog(parent, valueCheckRequired);
-    }
+		return super.showSetupDialog(parent, valueCheckRequired);
+	}
 
 }
